requirements.txt
# Core ML
torch
accelerate>=0.26.0

# Hugging Face stack
transformers
setfit
sentence-transformers
datasets
huggingface-hub>=0.34.0,<1.0

# Data & ML
pandas
numpy
scikit-learn

# Visualization
matplotlib
plotly

# App
streamlit
python-dotenv

# Keras compatibility (IMPORTANT)
tf-keras
